+++ CHANGELOG.md
@@ -8,12 +8,16 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/),
 and this project adheres to [Semantic Versioning](https://semver.org/).
 
 ## [Unreleased]
+### Added
+- `deno run <script>` accepts several scripts at the same time.
+
 ### Fixed
 - Improved the performance of loading page layouts by using the `Source` cache.
 - Improved error reporting.
 - `search.pages()` with no arguments
   returns pages with other extensions than `.html`.
 - Updated `nunjucks`.
+- Clear the cache before build. This allows to run several builds in the same script.
 
 ## [0.22.1] - 2021-06-05
 ### Fixed